Split up big file.
